Created Custom Buttons and added total price
Created custom buttons with a prop that will decide the color of the
button.  On the styles sheet, prop is taken in and thrown into an object
where a specific color hex will be used.  Also added total price to the
Order Summary so the customer can view the purchase price.