Order and Checkout components and Routing
Added routing to the application.  Also created a checkout component
that shows the user the current order and the user is able to fill out
a basic form.  The form is not currently connected to firebase.  The
Orders component shows the user the past orders the the user has sent
in.