Persistent Auth
Fixed routing so when a user creates a burger, it routes to the
signin/signup component if they aren't logged in.  After log in, they
will be routed to the checkout component.  Also, if the user logs in
and exits the program and comes back within the expiration time, the
user will still be logged in.